+++ system/src/Grav/Common/Twig/TwigExtension.php
@@ -149,6 +149,16 @@ class TwigExtension extends \Twig_Extension
     }
 
     /**
+     * @return array
+     */
+    public function getTokenParsers()
+    {
+        return [
+            new TokenParserTry(),
+        ];
+    }
+
+    /**
      * Filters field name by changing dot notation into array notation.
      *
      * @param  string $str