+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Added `Grav\Framework\Object` classes for creating collections of objects
     * Added `Grav\Framework\Page` interfaces
     * Added `|nicenumber` Twig filter
+    * Added `{% try %} ... {% catch %} Error: {{ e.message }} {% endcatch %}` tag to allow basic exception handling inside Twig
     * Deprecated GravTrait
 1. [](#improved)
     * Make it possible to include debug bar also into non-HTML responses